Measured Liquids only does tenths of fluid ounces
I believe I can redefine the units used, which honestly we could do
worse than working in metric and converting but it sounds like a huge
pain.

Much like it is IRL, especially with the British Teaspoon being off from
our teaspoon, it's a huge friggin' pain though practically most recipes
are forgiving enough that it won't kill it, it's just...annoying.
ANNOYING.