Add salt, mixing bowl
Now I need to figure out how to properly track mixtures.

The Measured Liquid extension is great so far but I don't think it quite
fits with what I want...it does have support for recipes, but it doesn't
have support for arbitrary quantity tracking, per se.

Let's go over the docs again. It might be that the combinatorial
explosion of Actually Modelling Cooking is impossible to easily, well,
model, but we'll give't the old college try eh?