Awful stand mixer setting code
Ok, the issue is I want...basically "an enum value that holds strings"
but Inform 7 parses "two" as the number 2 in the declaration, and
so we can't have "0, stir, 2" because then it's a kind of both
quantitative and qualitative values and it explodes.

And now I'm trying to hack around that. It ain't great.