...extremely cursed item creation issues.
OK SO.

When I reassign the various names of _Ingredient on the swapping of the
ingredient_info, I do so by creating a property of _Ingredient,
info_name.

I then allow the player to type 'x [info_name]' by using this line:
`Understand the info_name property as describing an _Ingredient;`
which, unfortunately, does *not* distinguish between two distinct
_Ingredients with the same info_name. Rather, it arbitrarily picks one!

That means if you have two different "water" objects, one of which is 4
gallons in a pot and one of which is 1 cup in a bowl, it doesn't ask you
which of the two to examine, it just picks one.

How vexing!